+++ lib/data/vs_node_manager.dart
@@ -12,6 +12,11 @@ class VSNodeManager {
   VSNodeManager({
     required List<dynamic> nodeBuilders,
     String? serializedNodes,
+    this.onNodesUpdate,
+
+    ///This function gets called when a node failed to deserialize due to a missing builder
+    ///
+    ///The nodeJSON in question is given as a parameter
     Function(Map nodeJSON)? onBuilderMissing,
 
     ///These nodes will not be part of [contextNodeBuilders]
@@ -34,14 +39,26 @@ class VSNodeManager {
   Map<String, dynamic> get nodeBuildersMap =>
       serializationManager.contextNodeBuilders;
 
-  ///Returns all output nodes in the current node data
-  Iterable<VSOutputNode> get getOutputNodes =>
-      _nodes.values.whereType<VSOutputNode>();
+  ///This function gets called whenever the nodes get updated
+  ///
+  ///Use it to run code synchronously
+  final void Function(
+    Map<String, VSNodeData> oldData,
+    Map<String, VSNodeData> newData,
+  )? onNodesUpdate;
 
   Map<String, VSNodeData> _nodes = {};
 
   ///Returns a copy of the current node data
   Map<String, VSNodeData> get nodes => Map.from(_nodes);
+  set nodes(Map<String, VSNodeData> value) {
+    onNodesUpdate?.call(_nodes, value);
+    _nodes = value;
+  }
+
+  ///Returns all output nodes in the current node data
+  Iterable<VSOutputNode> get getOutputNodes =>
+      _nodes.values.whereType<VSOutputNode>();
 
   ///Calls serializationManager.serializeNodes with the current node data and returns a String
   String serializeNodes() {
@@ -50,20 +67,7 @@ class VSNodeManager {
 
   ///Loades nodes from string and replaces current nodes
   void loadSerializedNodes(String serializedNodes) {
-    final deserializedNodes =
-        serializationManager.deserializeNodes(serializedNodes);
-
-    removeNodes(
-      nodes.keys
-          .toSet()
-          .difference(deserializedNodes.keys.toSet())
-          .map((e) => nodes[e]!)
-          .toList(),
-    );
-
-    updateOrCreateNodes(
-      deserializedNodes.values.toList(),
-    );
+    nodes = serializationManager.deserializeNodes(serializedNodes);
   }
 
   ///Updates or Creates a nodes
@@ -71,7 +75,7 @@ class VSNodeManager {
     for (final node in nodeDatas) {
       _nodes[node.id] = node;
     }
-    _nodes = Map.from(_nodes);
+    nodes = Map.from(_nodes);
   }
 
   ///Removes multiple nodes and clears all references
@@ -79,7 +83,7 @@ class VSNodeManager {
     for (final node in nodeDatas) {
       _nodes.remove(node.id);
     }
-    _nodes = Map.from(_nodes);
+    nodes = Map.from(_nodes);
     for (final node in _nodes.values) {
       for (final input in node.inputData) {
         if (nodeDatas.contains(input.connectedInterface?.nodeData)) {
@@ -91,6 +95,6 @@ class VSNodeManager {
 
   ///Cleares all nodes
   void clearNodes() async {
-    _nodes = {};
+    nodes = {};
   }
 }