+++ lib/vs_node_view.dart
@@ -13,6 +13,7 @@ export 'data/standard_interfaces/vs_num_interface.dart'
     show VSNumInputData, VSNumOutputData;
 export 'data/standard_interfaces/vs_string_interface.dart'
     show VSStringInputData, VSStringOutputData;
+export 'data/vs_history_manager.dart' show VSHistoryManger;
 //Data
 export 'data/vs_interface.dart' show VSInputData, VSOutputData;
 export 'data/vs_node_data.dart' show VSNodeData;
@@ -27,3 +28,5 @@ export 'special_nodes/vs_widget_node.dart' show VSWidgetNode;
 export 'widgets/interactive_vs_node_view.dart' show InteractiveVSNodeView;
 export 'widgets/vs_node.dart' show VSNode;
 export 'widgets/vs_node_view.dart' show VSNodeView;
+export 'widgets/inherited_node_data_provider.dart'
+    show InheritedNodeDataProvider;