+++ lib/widgets/vs_context_menu.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:vs_node_view/data/vs_node_data_provider.dart';
+import 'package:vs_node_view/widgets/inherited_node_data_provider.dart';
 
 class VSContextMenu extends StatefulWidget {
   ///Base context menu
@@ -61,7 +60,7 @@ class _VSContextMenuState extends State<VSContextMenu> {
       } else {
         widgets.add(TextButton(
           onPressed: () {
-            final dataProvider = context.read<VSNodeDataProvider>();
+            final dataProvider = InheritedNodeDataProvider.of(context).provider;
             dataProvider.createNodeFromContext(entry.value);
             dataProvider.closeContextMenu();
           },