+++ lib/widgets/vs_node_input.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
 import 'package:vs_node_view/common.dart';
 import 'package:vs_node_view/data/vs_interface.dart';
-import 'package:vs_node_view/data/vs_node_data_provider.dart';
 import 'package:vs_node_view/special_nodes/vs_list_node.dart';
+import 'package:vs_node_view/widgets/inherited_node_data_provider.dart';
 
 class VSNodeInput extends StatefulWidget {
   ///Base node input widget
@@ -37,7 +36,8 @@ class _VSNodeInputState extends State<VSNodeInput> {
   void didUpdateWidget(covariant VSNodeInput oldWidget) {
     super.didUpdateWidget(oldWidget);
 
-    if (widget.data.nodeData is VSListNode) {
+    if (widget.data.widgetOffset == null ||
+        widget.data.nodeData is VSListNode) {
       updateRenderBox();
     }
   }
@@ -54,7 +54,7 @@ class _VSNodeInputState extends State<VSNodeInput> {
 
   void updateConnectedNode(VSOutputData? data) {
     widget.data.connectedInterface = data;
-    context.read<VSNodeDataProvider>().updateOrCreateNodes(
+    InheritedNodeDataProvider.of(context).provider.updateOrCreateNodes(
       [widget.data.nodeData!],
     );
   }