+++ lib/widgets/vs_node_output.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
 import 'package:vs_node_view/common.dart';
 import 'package:vs_node_view/data/vs_interface.dart';
-import 'package:vs_node_view/data/vs_node_data_provider.dart';
 import 'package:vs_node_view/special_nodes/vs_list_node.dart';
 import 'package:vs_node_view/special_nodes/vs_widget_node.dart';
+import 'package:vs_node_view/widgets/inherited_node_data_provider.dart';
 import 'package:vs_node_view/widgets/line_drawer/gradiant_line_drawer.dart';
 
 class VSNodeOutput extends StatefulWidget {
@@ -40,7 +39,8 @@ class _VSNodeOutputState extends State<VSNodeOutput> {
   void didUpdateWidget(covariant VSNodeOutput oldWidget) {
     super.didUpdateWidget(oldWidget);
 
-    if (widget.data.nodeData is VSListNode) {
+    if (widget.data.widgetOffset == null ||
+        widget.data.nodeData is VSListNode) {
       updateRenderBox();
     }
   }
@@ -84,7 +84,7 @@ class _VSNodeOutputState extends State<VSNodeOutput> {
               dragPos = null;
             }),
             onDraggableCanceled: (velocity, offset) {
-              context.read<VSNodeDataProvider>().openContextMenu(
+              InheritedNodeDataProvider.of(context).provider.openContextMenu(
                     position: offset,
                     outputData: widget.data,
                   );